Port remap fix
Now connections between VMs are one port per connection. The ports
are exchanged via socket.io and a handshake + stun tries its best to
account for immediate port remaps, and the connection is established
if handshake is successful. If port remap happens mid-connection,
The XDP filter divert the packet to userspace where it updates the
ports.

We are also using a hashmap rather than iterating over a list in
userspace.